% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.table.performance.R
\name{as.data.table.performance}
\alias{as.data.table.performance}
\title{Convert Performance Objet to data.table}
\usage{
\method{as.data.table}{performance}(x, type = "metric", format = NULL, keep.rownames = FALSE, ...)
}
\arguments{
\item{x}{object of class \code{"performance"}.}

\item{type}{[character] either \code{"metric"} to extract AUC/brier score or \code{"prediction"} to extract predictions.}

\item{format}{[character] should the result be outcome in the long format (\code{"long"}) or in the wide format (\code{"wide"}).
Note relevant when using \code{type="metric"}.}

\item{keep.rownames}{Not used. For compatibility with the generic method.}

\item{...}{Not used. For compatibility with the generic method.}
}
\description{
Extract the AUC/brier score values or the prediction into a data.table format.
}
