% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.S4BuyseTest.R
\name{autoplot.S4BuyseTest}
\alias{autoplot.S4BuyseTest}
\title{Graphical Display for GPC}
\usage{
\method{autoplot}{S4BuyseTest}(
  object,
  type = "hist",
  strata = "global",
  endpoint = NULL,
  label.strata = NULL,
  label.endpoint = NULL,
  color = c("#7CAE00", "#F8766D", "#C77CFF", "#00BFC4"),
  ...
)
}
\arguments{
\item{object}{an \R object of class \code{\linkS4class{S4BuyseTest}}, i.e., output of \code{\link{BuyseTest}}}

\item{type}{[character] type of plot: histogram (\code{"hist"}), pie chart (\code{"pie"}), or nested pie charts (\code{"racetrack"}).}

\item{strata}{[character vector] strata(s) relative to which the percentage should be displayed.}

\item{endpoint}{[character vector] endpoint(s) relative to which the percentage should be displayed.}

\item{label.strata}{[character vector] new labels for the strata levels. Should match the length of argument \code{strata}.}

\item{label.endpoint}{[character vector] new labels for the endpoints. Should match the length of argument \code{endpoint}.}

\item{color}{[character vector] colors used to display the percentages for each type of pair.}

\item{...}{not used, for compatibility with the generic function.}
}
\value{
a ggplot object.
}
\description{
Graphical display of the percentage of favorable, unfavorable, neutral, and uninformative pairs per endpoint.
}
\keyword{hplot}
