% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuyseRes-object.R
\docType{class}
\name{BuyseRes-class}
\alias{BuyseRes-class}
\alias{BuyseRes}
\title{Class "BuyseRes" (output of BuyseTest)}
\description{
A \code{\link{BuyseTest}} output is reported in a \code{BuyseRes} object.
}
\details{
slots
}
\section{Slots}{

\describe{
\item{\code{count_favorable}}{[matrix] the probability for a random pair to be favorable for each strata (in rows) and each endpoint (in columns).}

\item{\code{count_unfavorable}}{[matrix] the probability for a random pair to be unfavorable for each strata (in rows) and each endpoint (in columns).}

\item{\code{count_neutral}}{[matrix] the probability for a random pair to be neutral for each strata (in rows) and each endpoint (in columns).}

\item{\code{count_uninf}}{[matrix] the probability for a random pair to be uninformative for each strata (in rows) and each endpoint (in columns).}

\item{\code{delta}}{[list of matrix] the chance of a better outcome
(net difference between the probability for a random pair to be favorable minus the probability to be unfavorable divided by the total number of pairs)
for each strata (in rows) and each endpoint (in columns).}

\item{\code{Delta}}{[list of numeric vector] the chance of a better outcome over all strata for each endpoint.}

\item{\code{delta_permutation}}{[list of array] the chance of a better outcome within each strata (first dimension),
over the endpoint (second dimension) and for each sample from the permutation test (third dimension).}

\item{\code{Delta_quantile}}{[list of matrix] the randomization test-based 2.5\% and the 97.5\% quantiles of the cumulative chance of a better outcome (in rows) for each endpoint (in columns).}

\item{\code{endpoint}}{[character vector] the name of the endpoint.}

\item{\code{index_neutralT}}{[integer vector] the index in the dataset of the treatment observations from remaining neutral pairs.}

\item{\code{index_neutralC}}{[integer vector] the index in the dataset of the control observations from remaining neutral pairs.}

\item{\code{index_uninfT}}{[integer vector] the index in the dataset of the treatment observations from remaining uninformative pairs.}

\item{\code{index_uninfC}}{[integer vector] the index in the dataset of the control observations from remaining uninformative pairs.}

\item{\code{levels.treatment}}{[character vector] the name of each group.}

\item{\code{n_pairs}}{[integer] the total number of pairs.}

\item{\code{n_permutation}}{[integer] the number of sucessful samples of the permutation test.}

\item{\code{p.value}}{[list of numeric vector] the p.value associated to the chance of a better outcome at each prioritized endpoint.}

\item{\code{strata}}{[character vector] the name of the strata.}

\item{\code{tableComparison}}{[list] detail of the result of each pairwise comparison relative to each endpoint.}

\item{\code{threshold}}{[numeric vector] the threshold associated to each endpoint.}
}}

\examples{
  n.Treatment_testBin <- 500
  n.Control_testBin <- 500
  prob.Treatment_testBin <- c(0.5,0.75)
  prob.Control_testBin <- c(0.5,0.25)
  
  set.seed(10)
  data_testBin <- data.frame(treatment=c(rep(1,n.Treatment_testBin),rep(0,n.Treatment_testBin)))
  data_testBin$endpoint1 <- c(rbinom(n.Treatment_testBin,size=1,prob=prob.Treatment_testBin[1]),
                              rbinom(n.Control_testBin,size=1,prob=prob.Control_testBin[1]))
  data_testBin$endpoint2 <- c(rbinom(n.Control_testBin,size=1,prob=prob.Treatment_testBin[2]),
                              rbinom(n.Control_testBin,size=1,prob=prob.Control_testBin[2]))
  data_testBin$strata <- rbinom(n.Treatment_testBin+n.Control_testBin,size=4,prob=0.5)
  
  #### no strata,
  BuyseTest_object <- BuyseTest(data=data_testBin,endpoint=c("endpoint1","endpoint2"),
                                treatment="treatment", type=c("bin","bin"), n.permutation = 0)
  
  class(BuyseTest_object)
  

}
\seealso{
\code{\link{BuyseTest}} for the function computing generalized pairwise comparisons. \cr
\code{\link{BuyseRes-summary}} for the summary of the BuyseTest function results
}
\keyword{BuyseRes-class}
\keyword{classes}
