% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-setGeneric.R, R/BuyseRes_get.R
\docType{methods}
\name{getIndividualScore}
\alias{getIndividualScore}
\alias{BuyseRes-getIndividualScore}
\alias{getIndividualScore}
\alias{getIndividualScore,BuyseRes-method}
\title{Extract the Score of Each Pair}
\usage{
getIndividualScore(object, ...)

\S4method{getIndividualScore}{BuyseRes}(object, endpoint = NULL,
  strata = NULL, rm.withinStrata = TRUE, rm.weight = FALSE,
  unlist = TRUE, trace = 1)
}
\arguments{
\item{object}{an \R object of class \code{\linkS4class{BuyseRes}}, i.e., output of \code{\link{BuyseTest}}}

\item{...}{not used. For compatibility with the generic method.}

\item{endpoint}{[integer/character vector] the endpoint for which the scores should be output.}

\item{strata}{[integer/character vector] the strata for which the scores should be output.}

\item{rm.withinStrata}{[logical] should the columns indicating the position of each member of the pair
within each treatment group be removed?}

\item{rm.weight}{[logical] should the column weight be remove from the output?}

\item{unlist}{[logical] should the structure of the output be simplified when possible?}

\item{trace}{[logical] should a message be printed to explain what happened
when the function returned \code{NULL}?}
}
\description{
Extract the score of each pair.
}
\details{
The maximal output (i.e. with all columns) contains for each endpoint, a data.table with:
\itemize{
\item \code{"strata"}: the name of the strata to which the pair belongs.
\item \code{"index.T"}: the index of the treatment observation in the pair relative to the original dataset.
\item \code{"index.C"}: the index of the control observation in the pair relative to the original dataset.
\item \code{"indexWithinStrata.T"}: the index of the treatment observation in the pair relative to the treatment group and the strata.
\item \code{"indexWithinStrata.C"}: the index of the control observation in the pair relative to the control group and the strata.
\item \code{"favorable"}: the probability that the endpoint is better in the treatment arm vs. in the control arm.
\item \code{"unfavorable"}: the probability that the endpoint is worse in the treatment arm vs. in the control arm.
\item \code{"neutral"}: the probability that the endpoint is no different in the treatment arm vs. in the control arm.
\item \code{"uninformative"}: the weight of the pair that cannot be attributed to favorable/unfavorable/neutral.
\item \code{"weight"}: the residual weight of the pair to be analysed at the current outcome. Each pair starts with a weight of 1.
}
Note that the \code{.T} and \code{.C} may change since they correspond of the label of the treatment and control arms.
}
\keyword{BuyseRes-method}
\keyword{get}
