% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuyseRes-summary.R
\docType{methods}
\name{BuyseRes-summary}
\alias{BuyseRes-summary}
\alias{summary,BuyseRes-method}
\title{Summary Method for Class "BuyseRes"}
\usage{
\S4method{summary}{BuyseRes}(object, print = TRUE, percentage = TRUE,
  statistic = NULL, conf.level = NULL, strata = if
  (length(object@level.strata) == 1) {     "global" } else {     NULL },
  digit = c(2, 4), ...)
}
\arguments{
\item{object}{output of \code{\link{BuyseTest}}}

\item{print}{[logical] Should the table be displayed?.}

\item{percentage}{[logical] Should the percentage of pairs of each type be displayed ? Otherwise the number of pairs is displayed.}

\item{statistic}{[character] the statistic summarizing the pairwise comparison:
\code{"netBenefit"} displays the net benefit, as described in Buyse (2010) and Peron et al. (2016)),
whereas \code{"winRatio"} displays the win ratio, as described in Wang et al. (2016).
Default value read from \code{BuyseTest.options()}.}

\item{conf.level}{[numeric] confidence level for the confidence intervals.
Default value read from \code{BuyseTest.options()}.}

\item{strata}{[character vector] the name of the strata to be displayed. Can also be \code{"global"} to display the average over all strata.}

\item{digit}{[integer vector] the number of digit to use for printing the counts and the delta.}

\item{...}{arguments to be passed to \code{confint}}
}
\description{
Summarize the results from the \code{\link{BuyseTest}} function.
}
\details{
\bold{Content of the output} \cr
The "results" table in the output show the result of the GPC at each endpoint, as well as its contribution to the global statistics.
More precisely, the column:
\itemize{
  \item \code{endpoint} lists the endpoints, by order of priority.
  \item \code{threshold} lists the threshold associated to each endpoint.
  \item \code{total} lists the total number of pairs (argument \code{percentage=FALSE}) or the percentage of pairs (argument \code{percentage=TRUE}) to be analyzed at the current priority.
  \item \code{favorable} lists the number of pairs (argument \code{percentage=FALSE}) or the percentage of pairs (argument \code{percentage=TRUE}) classified in favor of the treatment at the current priority.
  \item \code{unfavorable} lists the number of pairs (argument \code{percentage=FALSE}) or the percentage of pairs (argument \code{percentage=TRUE}) classified in favor of the control at the current priority.
  \item \code{neutral} lists the number of pairs (argument \code{percentage=FALSE}) or the percentage of pairs (argument \code{percentage=TRUE}) classified as neutral at the current priority.
  \item \code{uninf} lists the number of pairs (argument \code{percentage=FALSE}) or the percentage of pairs (argument \code{percentage=TRUE}) that could not be classified at the current priority (due to missing values/censoring).
  \item \code{delta} lists the value of the statistic (i.e. net benefit or win ratio) computed on the pairs analyzed at the current priority only.
  \item \code{Delta} lists the value of the statistic (i.e. net benefit or win ratio) computed on all the pairs analyzed up to the current priority.
  \item \code{CI} Confidence interval for the value of \code{Delta} (performed independently at each priority, no adjustment for multiple comparison).
  \item \code{p.value} p-value for the test \code{Delta=0} (performed independently at each priority, no adjustment for multiple comparison).
  \item \code{resampling} number of samples used to compute the confidence intervals or p-values from permutations or bootstrap samples.
Some bootstrap samples can be discarded if, for example, they did not lead to sample any case or control.
}
 
\bold{Statistical inference} \cr
When the interest is in obtaining p-values, we recommand the use of a permutation test.
However, when using a permutation test confidence intervals are not displayed in the summary.
This is because there is no (to the best of our knowledge) straightforward way to obtain good confidence intervals with permutations. 
An easy way consist in using the quantiles of the permutation distribution and then shift by the point estimate of the statistic.
This is what is output by \code{confint} method.
However this approach leads to a much too high coverage when the null hypothesis is false.
The limits of the confidence interval can also end up being outside of the interval of definition of the statistic
(e.g. outside [-1,1] for the proportion in favor of treatment).
Therefore, for obtaining confidence intervals, we recommand the boostrap method or the u-statistic method.

\bold{Win ratio} \cr
For the win ratio, the proposed implementation enables the use of thresholds and endpoints that are not time to events
as well as the correction proposed in Peron et al. (2016) to account for censoring. 
These development have not been examined by Wang et al. (2016), or in other papers (to the best of our knowledge).
They are only provided here by implementation convenience.
}
\examples{
library(data.table)

dt <- simBuyseTest(1e2, n.strata = 3)

 \dontrun{
 BT <- BuyseTest(Treatment ~ TTE(eventtime, censoring = status) + Bin(toxicity), data=dt)
 }
 \dontshow{
 BT <- BuyseTest(Treatment ~ TTE(eventtime, censoring = status) + Bin(toxicity), data=dt, n.resampling = 10, trace = 0)
 }
 summary(BT)
 summary(BT, percentage = FALSE)
 summary(BT, statistic = "winRatio")

}
\seealso{
\code{\link{BuyseTest}} for performing a generalized pairwise comparison. \cr
  \code{\link{BuyseRes-class}} for a presentation of the \code{BuyseRes} object.
}
\keyword{BuyseRes-method}
\keyword{summary}
