% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuyseSim-object.R
\docType{class}
\name{BuyseSim-class}
\alias{BuyseSim-class}
\alias{BuyseSim}
\title{Class "BuyseSim" (output of BuyseTest)}
\description{
A \code{\link{powerBuyseTest}} output is reported in a \code{BuyseSim} object.
}
\seealso{
\code{\link{powerBuyseTest}} for the function computing generalized pairwise comparisons. \cr
\code{\link{BuyseSim-summary}} for the summary of the BuyseTest function results
}
\keyword{BuyseSim-class}
\keyword{classes}
