% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuyseSim_show.R
\docType{methods}
\name{BuyseSim-show}
\alias{BuyseSim-show}
\alias{show,BuyseSim-method}
\title{Show Method for Class "BuyseSim"}
\usage{
\S4method{show}{BuyseSim}(object)
}
\arguments{
\item{object}{an \R object of class \code{\linkS4class{BuyseSim}}, i.e., output of \code{\link{BuyseTest}}}
}
\description{
Display the main results stored in a \code{\link{BuyseSim}} object.
}
\seealso{
\code{\link{BuyseTest}} for performing a generalized pairwise comparison. \cr
  \code{\link{BuyseSim-summary}} for a more detailed presentation of the \code{BuyseSim} object.
}
\keyword{BuyseSim-method}
\keyword{summary}
