% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuyseTest-package.R
\docType{package}
\name{BuyseTest-package}
\alias{BuyseTest-package}
\title{BuyseTest package: Generalized Pairwise Comparisons}
\references{
Methodological references: \cr
Marc Buyse (2010) \bold{Generalized pairwise comparisons of prioritized endpoints in the two-sample problem}. \emph{Statistics in Medicine} 29:3245-3257. \cr
J. Peron, M. Buyse, B. Ozenne, L. Roche and P. Roy (2016). \bold{An extension of generalized pairwise comparisons for prioritized outcomes in the presence of censoring}. Statistical Methods in Medical Research. \cr

\cr \cr
Example of application: \cr 
J. Peron, P Roy, K Ding, W R Parulekar, L Roche, M Buyse (2015). \bold{Assessing the benefit-risk of new treatments using generalised pairwise comparisons: the case of erlotinib in pancreatic cancer}. \emph{British journal of cancer} 112:(6)971-976.  \cr
J. Peron, M. Buyse, B. Ozenne, L. Roche and P. Roy (2016). \bold{The Net Chance of a Longer Survival as a Patient-Oriented Measure of Treatment Benefit in Randomized Clinical Trials}. \emph{JAMA Oncology} 2(7):901-5. \cr
}

