% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuyseTest.options-object.R
\docType{class}
\name{BuyseTest.option-class}
\alias{BuyseTest.option-class}
\title{Class "BuyseTest.option" (global setting for the BuyseTest package)}
\arguments{
\item{conf.level}{the confidence level of the confidence interval}

\item{keep.permutation}{should the result of each sample from the permutation test be stored in BuyseRes objects?}
}
\description{
Class defining the global settings for the BuyseTest package.
}
\seealso{
\code{\link{BuyseTest.option}} to select or update global settings.
}
\keyword{BuyseTest.option-class}
\keyword{classes}
\keyword{options}
