% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuyseTest.options-object.R
\docType{methods}
\name{BuyseTest.option-methods}
\alias{BuyseTest.option-methods}
\alias{alloc,BuyseTest.option-method}
\alias{select,BuyseTest.option-method}
\alias{select,BuyseTest.option-method}
\title{Methods for the class "BuyseTest.option"}
\usage{
\S4method{alloc}{BuyseTest.option}(object, field)

\S4method{select}{BuyseTest.option}(object, name.field)
}
\arguments{
\item{object}{an object of class \code{BuyseTest.option}.}

\item{field}{a \code{list} named with the name of the fields to update and containing the values to assign to these fields}

\item{name.field}{a \code{character vector} containing the names of the field to be selected.}
}
\description{
Methods to update or select global settings
}
