% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuyseTest.option.R
\name{BuyseTest.option}
\alias{BuyseTest.option}
\title{Global options for BuyseTest package}
\usage{
BuyseTest.option(..., reinitialise = FALSE)
}
\arguments{
\item{...}{options to be selected or updated}

\item{reinitialise}{should all the global parameters be set to their default value}
}
\description{
Update or select global options for the BuyseTest package.
}
\details{
It only affects the \code{\link{BuyseTest}} function
}
\examples{
 
## see all global parameters
BuyseTest.option()

## see some of the global parameters
BuyseTest.option("n.permutation", "trace")

## update some of the global parameters
BuyseTest.option(n.permutation = 10, trace = 1)
BuyseTest.option("n.permutation", "trace")

## reinitialise all global parameters
BuyseTest.option(reinitialise = TRUE)
}
