% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-BuyseTest.options.R
\name{BuyseTest.options-methods}
\alias{BuyseTest.options-methods}
\alias{alloc,BuyseTest.options-method}
\alias{select,BuyseTest.options-method}
\title{Methods for the class "BuyseTest.options"}
\usage{
\S4method{alloc}{BuyseTest.options}(object, field)

\S4method{select}{BuyseTest.options}(object, name.field)
}
\arguments{
\item{object}{an object of class \code{BuyseTest.options}.}

\item{field}{a \code{list} named with the name of the fields to update and containing the values to assign to these fields}

\item{name.field}{a \code{character vector} containing the names of the field to be selected.}
}
\description{
Methods to update or select global settings
}
