% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuyseTTEM.R
\name{BuyseTTEM}
\alias{BuyseTTEM}
\alias{BuyseTTEM.formula}
\alias{BuyseTTEM.prodlim}
\alias{BuyseTTEM.survreg}
\alias{BuyseTTEM.BuyseTTEM}
\title{Time to Event Model}
\usage{
BuyseTTEM(object, ...)

\method{BuyseTTEM}{formula}(object, treatment, iid, iid.surv = "exp", ...)

\method{BuyseTTEM}{prodlim}(object, treatment, iid, iid.surv = "exp", ...)

\method{BuyseTTEM}{survreg}(object, treatment, n.grid = NULL, iid, ...)

\method{BuyseTTEM}{BuyseTTEM}(object, ...)
}
\arguments{
\item{object}{time to event model.}

\item{...}{For compatibility with the generic method.}

\item{treatment}{[list or character] List containing the value of the treatment variable for each individual or name of the treatment variable.
The former is necessary when the survival model do not dependent on treatment.}

\item{iid}{[logical] Should the iid decomposition of the predictions be output.}

\item{iid.surv}{[character] Estimator of the survival used when computing the influence function.
Can be the product limit estimator (\code{"prodlim"}) or an exponential approximation (\code{"exp"}, same as in \code{riskRegression::predictCoxPL}).}

\item{n.grid}{[integer, >0] Number of timepoints used to discretize the time scale. Not relevant for prodlim objects.}
}
\value{
An S3 object of class \code{BuyseTTEM}.
}
\description{
Pre-compute quantities of a time to event model useful for predictions.
Only does something for prodlim objects.
}
\examples{
library(prodlim)
library(survival)
library(data.table)

tau <- seq(0,3,length.out=10)

#######################
#### survival case ####
#######################

set.seed(10)
df.data <- simBuyseTest(1e2, n.strata = 2)
df.Ta <- data.frame(treatment = "T", strata = "a")
df.Ca <- data.frame(treatment = "C", strata = "a")

#### Non-parametric model (stratified Kapan-Meier) ####

## estimate
eKM.prodlim <- prodlim(Hist(eventtime,status)~treatment+strata, data = df.data)
## plot(eKM.prodlim)

## convert to TTEM object
eKM.TTEM <- BuyseTTEM(eKM.prodlim, treatment = "treatment", iid = TRUE)

## extract jump times
jumpKM.Ta <- predict(eKM.TTEM, type = "jump", treatment = "T", strata = "a")
jumpKM.Ta
## extract last observation time
last.timeKM.Ta <- predict(eKM.TTEM, type = "last.time", treatment = "T", strata = "a")
last.timeKM.Ta
## extract survival at last observation time
last.survKM.Ta <- predict(eKM.TTEM, type = "last", treatment = "T", strata = "a")
last.survKM.Ta
## extract jump size
changeKM.Ta <- predict(eKM.TTEM, type = "change", treatment = "T", strata = "a")
last.survKM.Ta-sum(changeKM.Ta$survival) ## sum to 1
## prediction
predict(eKM.TTEM, time = tau, treatment = "T", strata = "a")
## same as prodlim: predict(eKM.prodlim, times = tau, newdata = df.Ta)
predict(eKM.TTEM, time = tau, treatment = "C", strata = "a")
## fix bug in prodlim: predict(eKM.prodlim, times = tau, newdata = df.Ca)
## survival should be 0 if the last observation is an event

#### Parametric model ####
dist <- "lognormal"

## estimate
eWei.survreg <- survreg(Surv(eventtime,status)~treatment+strata, data = df.data,
                        dist = dist)
## gridCDF <- seq(.01,.99,by=.01)
## plot(x = predict(eWei.survreg, newdata = df.Ta, type = "quantile", p = gridCDF,
##                  y = rev(gridCDF), col = "red", type = "l")
## points(x = predict(eWei.survreg, newdata = df.Ca, type = "quantile", p = gridCDF,
##                  y = rev(gridCDF), col = "blue", type = "l")

## conver to TTEM object
eWei.TTEM <- BuyseTTEM(eWei.survreg, treatment = "treatment", iid = TRUE, n.grid = 1e3)

## extract jump times
jumpWei.Ta <- predict(eWei.TTEM, type = "jump", treatment = "T", strata = "a")
jumpWei.Ta
## extract last observation time
last.timeWei.Ta <- predict(eWei.TTEM, type = "last.time", treatment = "T", strata = "a")
last.timeWei.Ta
## extract survival at last observation time
last.survWei.Ta <- predict(eWei.TTEM, type = "last", treatment = "T", strata = "a")
last.survWei.Ta
## extract jump size
changeWei.Ta <- predict(eWei.TTEM, type = "change", treatment = "T", strata = "a")
last.survWei.Ta-sum(changeWei.Ta$survival) ## sum to 1
## prediction
predWei.Ta <- predict(eWei.TTEM, time = tau, treatment = "T", strata = "a")
predWei.Ta
## survreg: predict(eWei.survreg, newdata = df.Ta, type = "quantile", p = 1-predWei.Ta$survival)-tau
## approximate retrieve tau, up to the grid resolution

#### competing risk case ####
df.dataCR <- copy(df.data)
df.dataCR$status <- rbinom(NROW(df.dataCR), prob = 0.5, size = 2)

eAJ.prodlim <- prodlim(Hist(eventtime,status)~treatment+strata, data = df.dataCR)
## plot(eAJ.prodlim)

eAJ.TTEM <- BuyseTTEM(eAJ.prodlim, treatment = "treatment", iid = TRUE)

## extract jump times
jumpAJ.Ta <- predict(eAJ.TTEM, type = "jump", treatment = "T", strata = "a")
jumpAJ.Ta
## extract last observation time
last.timeAJ.Ta <- predict(eAJ.TTEM, type = "last.time", treatment = "T", strata = "a")
last.timeAJ.Ta
## extract survival at last observation time
last.cifAJ.Ta <- predict(eAJ.TTEM, type = "last", treatment = "T", strata = "a")
last.cifAJ.Ta
## extract jump size
changeAJ.Ta <- predict(eAJ.TTEM, type = "change", treatment = "T", strata = "a")
last.cifAJ.Ta-sum(changeAJ.Ta$cif) ## sum to 0
## prediction
predict(eAJ.TTEM, time = tau, treatment = "T", strata = "a")
## same as prodlim: predict(eAJ.prodlim, times = tau, newdata = df.Ta, cause = 1)

predict(eAJ.TTEM, time = tau, treatment = "C", strata = "a")
## same as prodlim: predict(eAJ.prodlim, times = tau, newdata = df.Ca, cause = 1)
}
\keyword{models}
