% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.R
\name{coef.BuyseTestAuc}
\alias{coef.BuyseTestAuc}
\title{Extract the AUC Value}
\usage{
\method{coef}{BuyseTestAuc}(object, ...)
}
\arguments{
\item{object}{object of class \code{BuyseTestAUC} (output of the \code{auc} function).}

\item{...}{not used. For compatibility with the generic function.}
}
\value{
Estimated value for the AUC (numeric).
}
\description{
Extract the AUC value.
}
