% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brier.R
\name{coef.BuyseTestBrier}
\alias{coef.BuyseTestBrier}
\title{Extract the Brier Score}
\usage{
\method{coef}{BuyseTestBrier}(object, ...)
}
\arguments{
\item{object}{object of class \code{BuyseTestBrier} (output of the \code{brier} function).}

\item{...}{not used. For compatibility with the generic function.}
}
\value{
Estimated value for Brier score (numeric).
}
\description{
Extract the Brier score.
}
\keyword{methods}
