% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brier.R
\name{confint.BuyseTestBrier}
\alias{confint.BuyseTestBrier}
\title{Extract the Brier Score with its Confidence Interval}
\usage{
\method{confint}{BuyseTestBrier}(object, ...)
}
\arguments{
\item{object}{object of class \code{BuyseTestBrier} (output of the \code{brier} function).}

\item{...}{not used. For compatibility with the generic function.}
}
\value{
Estimated value for the brier score, its standard error, the lower and upper bound of the confidence interval and the p-value.
}
\description{
Extract the Brier score with its Confidence Interval and possibly a p-value.
}
\keyword{methods}
