% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-data.R
\docType{data}
\name{EB}
\alias{EB}
\title{Rare disease trial}
\format{
An object of class \code{data.frame} with 30 rows and 7 columns.
}
\usage{
data(CHARM, package = "BuyseTest")
}
\description{
16 pediatric subjects suffering from a aare skin disease (epidermolysis bullosa simplex) treated with placebo or diacerin cream in a longitudinal cross-over trial (14 paired).

\itemize{
\item \code{Id}: study participant.
\item \code{Time}: end of the 4 weeks treatment period (t4) or 12 weeks after treatment ended (t12)
\item \code{treatment}: treatment arm: diacerin cream (V) or placebo (P).
\item \code{StdDiffCount}: blister uncertainty
\item \code{Bin}: indicator of 40% reduction from baseline in the number of blisters within the treated areas.
\item \code{DiffQoL}: Change in quality of life since baseline.
\item \code{period}: same as time but coded as numeric: 1 for t4 and 2 for t12
}
}
\references{
Wally et al. "Diacerein orphan drug development for epidermolysis bullosa simplex: A phase 2/3 randomized, placebo-controlled, double-blind clinical trial". Journal of American Academy of Dermatology (2018) 78(5):892-901. https://doi.org/10.1016/j.jaad.2018.01.019.
}
\keyword{datasets}
