% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.colCumSum_cpp}
\alias{.colCumSum_cpp}
\title{Column-wise cumulative sum}
\usage{
.colCumSum_cpp(X)
}
\arguments{
\item{X}{A matrix.}
}
\value{
A matrix of same size as x.
}
\description{
Fast computation of apply(x,2,cumsum)
}
\keyword{utilities}
