% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.colScale_cpp}
\alias{.colScale_cpp}
\title{Divide by a vector of values in each column}
\usage{
.colScale_cpp(X, scale)
}
\arguments{
\item{X}{A matrix.}

\item{scale}{A vector with length the number of rows of X .}
}
\value{
A matrix of same size as x.
}
\description{
Fast computation of sweep(X, FUN = "/", STATS = scale, MARGIN = 1)
}
\keyword{utilities}
