% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.logit.R
\name{.information.logit}
\alias{.information.logit}
\title{Information for Logistic Regressions}
\usage{
.information.logit(object, indiv, center)
}
\arguments{
\item{object}{a glm object corresponding to a logistic regression.}

\item{indiv}{[logical] should the individual contribution be output instead of the total information?}

\item{center}{[logical] should the individual contribution be centered around the average?}
}
\description{
Compute the information (i.e. opposit of the expectation of the second derivative of the log-likelihood) for logistic regressions.
}
\keyword{internal}
