% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.rowCumProd_cpp}
\alias{.rowCumProd_cpp}
\title{Apply cumprod in each row}
\usage{
.rowCumProd_cpp(X)
}
\arguments{
\item{X}{A matrix.}
}
\value{
A matrix of same size as x.
}
\description{
Fast computation of t(apply(x,1,cumprod))
}
\keyword{utilities}
