% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.rowCumSum_cpp}
\alias{.rowCumSum_cpp}
\title{Row-wise cumulative sum}
\usage{
.rowCumSum_cpp(X)
}
\arguments{
\item{X}{A matrix.}
}
\value{
A matrix of same size as x.
}
\description{
Fast computation of apply(x,1,cumsum)
}
\keyword{utilities}
