% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.rowScale_cpp}
\alias{.rowScale_cpp}
\title{Dividy by a vector of values in each row}
\usage{
.rowScale_cpp(X, scale)
}
\arguments{
\item{X}{A matrix.}

\item{scale}{A vector with length the number of columns of X.}
}
\value{
A matrix of same size as x.
}
\description{
Fast computation of sweep(X, FUN = "/", STATS = center, MARGIN = 2)
}
