% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.logit.R
\name{.score.logit}
\alias{.score.logit}
\title{Score for Logistic Regressions}
\usage{
.score.logit(object, indiv)
}
\arguments{
\item{object}{a glm object corresponding to a logistic regression.}

\item{indiv}{[logical] should the individual contribution be output instead of the total score?}
}
\description{
Compute the first derivative of the log-likelihood IPCW logistic regressions.
}
\keyword{internal}
