% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.logit.R
\name{.vcov.logit}
\alias{.vcov.logit}
\title{Variance-covariance matrix for Logistic Regressions}
\usage{
.vcov.logit(object, indiv, center)
}
\arguments{
\item{object}{a glm object corresponding to a logistic regression.}

\item{indiv}{[logical] should the individual contribution be output instead of the total variance-covariance?}

\item{center}{[logical] should the individual contribution be centered around the average?}
}
\description{
Compute the variance covariance matrix (i.e. inverse of the information) for logistic regressions.
}
\keyword{internal}
