% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalBuyseTest.R
\name{evalBuyseTest}
\alias{evalBuyseTest}
\title{Performing simulation studies with BuyseTest}
\usage{
evalBuyseTest(sim, sample.size, sample.sizeC = NULL, sample.sizeT = NULL,
  n.rep, cpus = 1, alternative = "two.sided", seed = NULL,
  conf.level = 0.95, trace = 1, ...)
}
\arguments{
\item{sim}{[function] take two arguments:
the sample size in the control group (\code{n.C}) and the sample size in the treatment group (\code{n.C})
and generate datasets. The datasets must be data.table objects.}

\item{sample.size}{the various sample sizes at which the simulation should be perform.
Disregarded if any of the arguments \code{sample.sizeC} or \code{sample.sizeT} are specified.}

\item{sample.sizeC}{the various sample sizes in the control group.}

\item{n.rep}{the number of simulations.}

\item{cpus}{[integer, >0] the number of CPU to use.
Only the permutation test can use parallel computation.
Default value read from \code{BuyseTest.options()}.}

\item{alternative}{[character] the alternative hypothesis.
Must be one of \code{"two.sided"}, \code{"greater"} or \code{"less"}.}

\item{seed}{[integer, >0] the seed to consider for the simulation study.}

\item{conf.level}{[numeric] confidence level for the confidence intervals.}

\item{trace}{[integer] should the execution of the function be traced?}

\item{...}{parameters from \code{BuyseTest}.}

\item{sample.sizeC}{the various sample sizes in the treatment group.}
}
\description{
Performs a simulation studies for several sample sizes.
Returns estimates, standard errors, confidence intervals and p.values.
}
\examples{

simFCT <- function(n.C, n.T){
    out <- data.table(Y=rnorm(n.C+n.T),
                      T=c(rep(1,n.C),rep(0,n.T))
                     )
return(out)
}

evalBuyseTest(sim = simFCT, sample.size = c(10,100), n.rep = 2,
              formula = T ~ cont(Y), method.inference = "asymptotic", trace = 4)


}
