% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-BuyseTest-get.R
\docType{methods}
\name{getCount}
\alias{getCount}
\alias{getCount,S4BuyseTest-method}
\title{Extract the Number of Favorable, Unfavorable, Neutral, Uninformative pairs}
\usage{
\S4method{getCount}{S4BuyseTest}(object, type)
}
\arguments{
\item{object}{an \R object of class \code{\linkS4class{S4BuyseTest}}, i.e., output of \code{\link{BuyseTest}}}

\item{type}{the type of pairs to be counted. Can be \code{"favorable"}, \code{"unfavorable"}, \code{neutral}, or \code{uninf}. Can also be \code{"all"} to select all of them.}
}
\value{
A \code{"vector"} containing the number of pairs
}
\description{
Extract the number of favorable, unfavorable, neutral, uninformative pairs.
}
\author{
Brice Ozenne
}
\keyword{S4BuyseTest-method}
\keyword{get}
