% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-BuyseTest-get.R
\docType{methods}
\name{getMatrixScore}
\alias{getMatrixScore}
\alias{getMatrixScore,S4BuyseTest-method}
\title{Extract the Score Matrix}
\usage{
\S4method{getMatrixScore}{S4BuyseTest}(
  object,
  endpoint = NULL,
  statistic = NULL,
  strata = NULL,
  cumulative = TRUE,
  within = FALSE,
  unlist = TRUE
)
}
\arguments{
\item{object}{an \R object of class \code{\linkS4class{S4BuyseTest}}, i.e., output of \code{\link{BuyseTest}}}

\item{endpoint}{[integer/character vector] the endpoint for which the scores should be output.}

\item{statistic}{[character] the statistic summarizing the pairwise comparison: \code{"netBenefit"}, \code{"favorable"}, \code{"unfavorable"}, \code{"neutral"}, \code{"uninf"}.}

\item{strata}{[character vector] the strata relative to which the score should be output.}

\item{cumulative}{[logical] should the scores be cumulated over endpoints?}

\item{within}{[logical] should within group pairwise comparisons be exported?
This requires to re-run GPC which can be time and memory consuming.}

\item{unlist}{[logical] should the structure of the output be simplified when possible?}
}
\description{
Extract all pairwise comparisons in a matrix format.
}
\examples{
library(data.table)
library(prodlim)
library(survival) ## import veteran

#### Gehan scoring rule ####
## non stratified GPC
e.Gehan <- BuyseTest(trt ~ tte(time, threshold = 20, status = "status"),
                     data = veteran, scoring.rule = "Gehan", keep.pairScore = TRUE)
Mscore.Gehan <- getMatrixScore(e.Gehan)
mean(Mscore.Gehan) - coef(e.Gehan) ## same
m <- NROW(Mscore.Gehan)
n <- NCOL(Mscore.Gehan)

var1.Gehan <- sum((colSums(Mscore.Gehan)/m - mean(Mscore.Gehan))^2)
var2.Gehan <- sum((rowSums(Mscore.Gehan)/n - mean(Mscore.Gehan))^2)
var1.Gehan/n^2 + var2.Gehan/m^2 - confint(e.Gehan)[,"se"]^2 ## same

MscoreFull.Gehan <- getMatrixScore(e.Gehan, within = TRUE)
table(MscoreFull.Gehan[1:m,m+(1:n)] - Mscore.Gehan) ## same
all(MscoreFull.Gehan[1:m,m+(1:n)] + t(MscoreFull.Gehan[m+(1:n),1:m])==0)

## stratified GPC
e.GehanS <- BuyseTest(trt ~ tte(time, threshold = 20, status = "status") + celltype,
                     data = veteran, scoring.rule = "Gehan", keep.pairScore = TRUE)
Mscore.GehanS <- getMatrixScore(e.GehanS)
Mscore.GehanS

Mscore.GehanS1 <- getMatrixScore(e.GehanS, strata = "adeno")
Mscore.GehanS1

MscoreFull.GehanS <- getMatrixScore(e.GehanS, within = TRUE)

#### Peron scoring rule ####
## non stratified GPC
e.Peron <- BuyseTest(trt ~ tte(time, threshold = 20, status = "status"),
                     data = veteran[veteran$celltype == "adeno",], keep.pairScore = TRUE)
Mscore.Peron <- getMatrixScore(e.Peron)
mean(Mscore.Peron) - coef(e.Peron) ## same
m <- NROW(Mscore.Peron)
n <- NCOL(Mscore.Peron)

var1.Peron <- sum((colSums(Mscore.Peron)/m - mean(Mscore.Peron))^2)
var2.Peron <- sum((rowSums(Mscore.Peron)/n - mean(Mscore.Peron))^2)
var1.Peron/n^2 + var2.Peron/m^2 - confint(e.Peron)[,"se"]^2 ## too low
## as ignore the uncertainty about the survival model

MscoreFull.Peron <- getMatrixScore(e.Peron, within = TRUE)
table(MscoreFull.Peron[1:m,m+(1:n)] - Mscore.Peron) ## same
all(MscoreFull.Peron[1:m,m+(1:n)] + t(MscoreFull.Peron[m+(1:n),1:m])==0)

MscoreFull.Peron[2,m+(1:n)]
Mscore.Peron[2,]

## stratified GPC
e.PeronS <- BuyseTest(trt ~ tte(time, threshold = 20, status = "status") + celltype,
                     data = veteran, keep.pairScore = TRUE)
Mscore.PeronS <- getMatrixScore(e.PeronS)
Mscore.PeronS

Mscore.PeronS1 <- getMatrixScore(e.PeronS, strata = "adeno")
Mscore.PeronS1

MscoreFull.PeronS <- getMatrixScore(e.PeronS, within = TRUE)
}
\author{
Brice Ozenne
}
\keyword{S4BuyseTest-method}
\keyword{get}
