% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.R
\name{iid.BuyseTestAuc}
\alias{iid.BuyseTestAuc}
\title{Extract the idd Decomposition for the AUC}
\usage{
\method{iid}{BuyseTestAuc}(object, ...)
}
\arguments{
\item{object}{object of class \code{BuyseTestAUC} (output of the \code{auc} function).}

\item{...}{not used. For compatibility with the generic function.}
}
\value{
A column vector.
}
\description{
Extract the iid decompotion relative to AUC estimate.
}
