% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brier.R
\name{iid.BuyseTestBrier}
\alias{iid.BuyseTestBrier}
\title{Extract the idd Decomposition for the Brier Score}
\usage{
\method{iid}{BuyseTestBrier}(x, ...)
}
\arguments{
\item{x}{object of class \code{BuyseTestBrier} (output of the \code{brier} function).}

\item{...}{not used. For compatibility with the generic function.}
}
\value{
A column vector.
}
\description{
Extract the iid decompotion relative to Brier score estimate.
}
\keyword{methods}
