% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuyseTest-computation.R
\name{internal-computation}
\alias{internal-computation}
\alias{calcPermutation}
\alias{calcCI}
\alias{calcCI}
\alias{warper_BTpermutation}
\title{internal functions for BuyseTest - computations}
\usage{
calcPermutation(envir)

calcCI(Delta, Delta_permutation, endpoint, D, alternative, alpha, n.permutation,
  cpus, trace)

warper_BTpermutation(x, envir)
}
\description{
Functions called by \code{\link{BuyseTest}} to perform calculations.
}
\details{
\code{calcPermutation} performs the permutation test in sequential or parallel mode. \cr
\code{calcCI} post process the results to obtain the p.value and the confidence interval. \cr
\code{warper_BTpermutation} resample the data for the cpp function. \cr
}
\keyword{BuyseTest}
\keyword{function}
\keyword{internal}
