% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuyseTest-print.R
\name{internal-print}
\alias{internal-print}
\alias{printGeneral}
\title{internal functions for BuyseTest - display}
\usage{
printGeneral(censoring, D, D.TTE, data, endpoint, level.strata, level.treatment,
  method, neutral.as.uninf, operator, strata, threshold, threshold.TTEM1,
  treatment, type, Wscheme, ...)
}
\description{
Functions called by \code{\link{BuyseTest}} to display the settings.
}
\keyword{internal}
