% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iid.logit.R
\name{predict.logit}
\alias{predict.logit}
\alias{.predict.glm}
\title{Predicted Probability with Influence Function}
\usage{
.predict.glm(object, newdata, level = NULL)
}
\arguments{
\item{object}{a logistic model.}

\item{newdata}{[data.frame] dataset containing}

\item{level}{[character] level of the outcome for which the probability should be computed.}
}
\description{
Compute the predicted probabilities from a logistic regression,
with their (robust) standard error,
and the corresponding influence function.
}
\keyword{internal}
