% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-BuysePower.R
\docType{class}
\name{S4BuysePower-class}
\alias{S4BuysePower-class}
\title{Class "S4BuysePower" (output of BuyseTest)}
\description{
A \code{\link{powerBuyseTest}} output is reported in a \code{S4BuysePower} object.
}
\seealso{
\code{\link{powerBuyseTest}} for the function computing generalized pairwise comparisons. \cr
\code{\link{S4BuysePower-summary}} for the summary of the BuyseTest function results
}
\author{
Brice Ozenne
}
\keyword{S4BuysePower-class}
\keyword{classes}
