% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-BuysePower-nobs.R
\docType{methods}
\name{S4BuysePower-nobs}
\alias{S4BuysePower-nobs}
\alias{nobs,S4BuysePower-method}
\title{Sample Size for Class "S4BuysePower"}
\usage{
\S4method{nobs}{S4BuysePower}(object, ...)
}
\arguments{
\item{object}{an \R object of class \code{S4BuysePower}, i.e., output of \code{\link{powerBuyseTest}}}

\item{...}{no used, for compatibility with the generic method.}
}
\value{
A data.frame with two colunms, one for each treatment group, and as many rows as sample sizes used for the simulation.
}
\description{
Display the sample size in each treatmnet arm as well as the number of pairs.
}
\author{
Brice Ozenne
}
\keyword{methods}
