% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-BuysePower-print.R
\docType{methods}
\name{S4BuysePower-print}
\alias{S4BuysePower-print}
\alias{print,S4BuysePower-method}
\title{Print Method for Class "S4BuysePower"}
\usage{
\S4method{print}{S4BuysePower}(x, ...)
}
\arguments{
\item{x}{an \R object of class \code{S4BuysePower}, i.e., output of \code{\link{powerBuyseTest}}}

\item{...}{additional arguments passed to the summary method.}
}
\value{
invisible table
}
\description{
Display the main results stored in a \code{S4BuysePower} object.
}
\seealso{
\code{\link{powerBuyseTest}} for performing power calculation based on GPC. \cr
  \code{\link{S4BuysePower-summary}} for a more detailed presentation of the \code{S4BuysePower} object.
}
\author{
Brice Ozenne
}
\keyword{print}
