% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-BuysePower-show.R
\docType{methods}
\name{S4BuysePower-show}
\alias{S4BuysePower-show}
\alias{show,S4BuysePower-method}
\title{Show Method for Class "S4BuysePower"}
\usage{
\S4method{show}{S4BuysePower}(object)
}
\arguments{
\item{object}{an \R object of class \code{S4BuysePower}, i.e., output of \code{\link{BuyseTest}}}
}
\description{
Display the main results stored in a \code{S4BuysePower} object.
}
\seealso{
\code{\link{BuyseTest}} for performing a generalized pairwise comparison. \cr
  \code{\link{S4BuysePower-summary}} for a more detailed presentation of the \code{S4BuysePower} object.
}
\author{
Brice Ozenne
}
\keyword{S4BuysePower-method}
\keyword{summary}
