% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-BuyseTest.R
\docType{class}
\name{S4BuyseTest-class}
\alias{S4BuyseTest-class}
\title{Class "S4BuyseTest" (output of BuyseTest)}
\description{
A \code{\link{BuyseTest}} output is reported in a \code{S4BuyseTest} object.
}
\seealso{
\code{\link{BuyseTest}} for the function computing generalized pairwise comparisons. \cr
\code{\link{S4BuyseTest-summary}} for the summary of the BuyseTest function results
}
\author{
Brice Ozenne
}
\keyword{S4BuyseTest-class}
\keyword{classes}
