% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-BuyseTest-nobs.R
\docType{methods}
\name{S4BuyseTest-nobs}
\alias{S4BuyseTest-nobs}
\alias{nobs,S4BuyseTest-method}
\title{Sample Size for Class "S4BuyseTest"}
\usage{
\S4method{nobs}{S4BuyseTest}(object, strata = FALSE, resampling = FALSE, simplify = TRUE, ...)
}
\arguments{
\item{object}{an \R object of class \code{S4BuyseTest}, i.e., output of \code{\link{BuyseTest}}}

\item{strata}{[character vector] the strata relative to which the number of pairs should be output.
Can also be \code{"global"} or \code{FALSE} to output the total number of pairs (i.e. across all strata),
or \code{TRUE} to output each strata-specific number of pairs.}

\item{resampling}{[logical] should the sample size of each bootstrap or permutation sample be output?}

\item{simplify}{[logical] should the result be coerced to the lowest possible dimension?}

\item{...}{no used, for compatibility with the generic method.}
}
\value{
A vector (when argument \code{strata} is \code{FALSE}) or a matrix (when argument \code{strata} is \code{TRUE}). In the latter case each line correspond to a strata.
}
\description{
Display the sample size in each treatment arm as well as the number of pairs.
}
\author{
Brice Ozenne
}
\keyword{methods}
