% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-BuyseTest-print.R
\docType{methods}
\name{S4BuyseTest-print}
\alias{S4BuyseTest-print}
\alias{print,S4BuyseTest-method}
\title{Print Method for Class "S4BuyseTest"}
\usage{
\S4method{print}{S4BuyseTest}(x, ...)
}
\arguments{
\item{x}{an \R object of class \code{S4BuyseTest}, i.e., output of \code{\link{BuyseTest}}}

\item{...}{additional arguments passed to the summary method.}
}
\description{
Display the main results stored in a \code{S4BuyseTest} object.
}
\seealso{
\code{\link{BuyseTest}} for performing a generalized pairwise comparison. \cr
  \code{\link{S4BuyseTest-summary}} for a more detailed presentation of the \code{S4BuyseTest} object.
}
\author{
Brice Ozenne
}
\keyword{print}
