% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-BuysePower-show.R, R/S4-BuyseTest-show.R
\docType{methods}
\name{S4BuysePower-show}
\alias{S4BuysePower-show}
\alias{show,S4BuysePower-method}
\alias{S4BuyseTest-show}
\alias{show,S4BuyseTest-method}
\title{Show Method for Class "S4BuysePower"}
\usage{
\S4method{show}{S4BuysePower}(object)

\S4method{show}{S4BuyseTest}(object)
}
\arguments{
\item{object}{an \R object of class \code{S4BuyseTest}, i.e., output of \code{\link{BuyseTest}}}
}
\value{
invisible \code{NULL}
}
\description{
Display the main results stored in a \code{S4BuysePower} object.

Display the main results stored in a \code{S4BuyseTest} object.
}
\seealso{
\code{\link{powerBuyseTest}} for performing power calculation based on GPC. \cr
  \code{\link{S4BuysePower-summary}} for a more detailed presentation of the \code{S4BuysePower} object.

\code{\link{BuyseTest}} for performing a generalized pairwise comparison. \cr
  \code{\link{S4BuyseTest-summary}} for a more detailed presentation of the \code{S4BuyseTest} object.
}
\author{
Brice Ozenne
}
\keyword{print}
