% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCT_simulBT.R
\name{Simulation function}
\alias{Simulation function}
\alias{simulBT}
\title{Simulation of data for the BuyseTest}
\usage{
simulBT(n.T, n.C = NULL, argsBin = list(), argsCont = list(),
  argsTTE = list(), n.strata = NULL, names.strata = NULL,
  format = "data.table")
}
\arguments{
\item{n.T}{number of patients in the treatment arm}

\item{n.C}{number of patients in the control arm}

\item{argsBin}{a list of arguments to be passed to simulBT_bin. They specify the distribution parameters of the binary endpoints.}

\item{argsCont}{a list of arguments to be passed to simulBT_continous. They specify the distribution parameters of the continuous endpoints.}

\item{argsTTE}{a list of arguments to be passed to simulBT_TTE. They specify the distribution parameters of the time to event endpoints.}

\item{n.strata}{number of strata. \code{NULL} indicates no strata.}

\item{names.strata}{name of the strata varaibles. Must have same length as \code{n.strata}.}

\item{format}{the format of the output. Can be "data.table", "data.frame" or "matrix".}
}
\description{
Simulate binary, continuous or time to event data, possibly with strata.
Outcomes are simulated independently of each other and independently of the strata variable.
}
\details{
This function is built upon the \code{lvm} and \code{sim} functions from the lava package.

Arguments in the list \code{argsBin}:
\itemize{
    \item\code{p.T} probability of event of each endpoint (binary endpoint, treatment group). \cr 
    \item\code{p.C} same as \code{p.T} but for the control group. \cr
    \item\code{name} names of the binary variables. \cr
}

Arguments in the list \code{argsCont}:
    \itemize{
    \item\code{mu.T} expected value of each endpoint (continuous endpoint, treatment group). \cr 
    \item\code{mu.C} same as \code{mu.C} but for the control group. \cr
    \item\code{sigma.T} standard deviation of the values of each endpoint (continuous endpoint, treatment group). \cr 
    \item\code{sigma.C} same as \code{sigma.T} but for the control group. \cr
    \item\code{name} names of the continous variables.
    }

Arguments in the list \code{argsTTE}:
    \itemize{
    \item\code{rates.T} hazard corresponding to each endpoint (time to event endpoint, treatment group). \cr 
    \item\code{rates.C} same as \code{rates.T} but for the control group. \cr
    \item\code{rates.Censoring} Censoring same as \code{rates.T} but for the censoring. \cr
    \item\code{name} names of the time to event variables. \cr
    \item\code{nameCensoring} names of the event type indicators. \cr
    }
}
\examples{
simulBT(1e3)

## only binary endpoints
simulBT(1e3, argsBin = list(p.T = c(3:5/10)), argsCont = NULL, argsTTE = NULL)

## only continous endpoints
simulBT(1e3, argsBin = NULL, argsCont = list(mu.T = c(3:5/10), sigma.T = rep(1,3)), argsTTE = NULL)

## only TTE endpoints
simulBT(1e3, argsBin = NULL, argsCont = NULL, 
        argsTTE = list(rates.T = c(3:5/10), rates.Censoring = rep(1,3)))
        
## one strata with 5 levels
simulBT(1e3, n.strata = 5)
simulBT(1e3, n.strata = 5, names.strata = "grade")

## several strata
simulBT(1e3, n.strata = c(2,4), names.strata = c("Gender","AgeCategory"))

}
\keyword{function}
\keyword{simulations}
