% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simCompetingRisks.R
\name{Simulation function}
\alias{Simulation function}
\alias{simCompetingRisks}
\title{Simulation of Gompertz competing risks data for the BuyseTest}
\usage{
simCompetingRisks(n.T, n.C, p.1C = NULL, v.1C, v.1T, v.2C, v.2T,
  sHR = NULL, b.1T = NULL, b.1C = NULL, b.2T = NULL, b.2C = NULL,
  cens.distrib = NULL, param.cens = NULL, latent = NULL)
}
\arguments{
\item{n.T}{[integer, >0] number of patients in the treatment arm}

\item{n.C}{[integer, >0] number of patients in the control arm}

\item{v.1C, v.1T, v.2C, v.2T}{shape parameters for distribution of time to event of interest in control/treatment (C/T) group and of time 
to competing event in control/treatment (C/T) group respectively}

\item{sHR}{[double, >0] pre-specified sub-distribution hazard ratio for event of interest}

\item{b.1C, b.1T, b.2C, b.2T}{rate parameters for distribution of time to event of interest in control/treatment (C/T) group and of time 
to competing event in control/treatment (C/T) group respectively. Can be NULL if and only if \code{(p.1C, sHR)} are provided.}

\item{cens.distrib}{[character] censoring distribution. Can be \code{"exponential"} for exponential censoring or \code{"uniform"} for
uniform censoring. NULL means no censoring.}

\item{param.cens}{[>0] parameter for censoring distribution. Should be a double for rate parameter of exponential censoring or a vector 
of doubles for lower and upper bounds of uniform censoring. NULL means no censoring}

\item{latent}{[logical] If \code{TRUE} also export the latent variables (e.g. true event times, event types and censoring times). NULL
sets this parameter to \code{FALSE}.}

\item{p1C}{[integer, >0] proportion of events of interest in the control group. Can be NULL if and only if \code{(b.1T, b.1C, b.2T, b.2C)}
are provided.}
}
\description{
Simulate Gompertz competing risks data with proportional (via prespecified sub-distribution hazard ratio) or
non-proportional sub-distribution hazards. A treatment variable with two groups (treatment and control) is created.
}
\details{
Arguments in the list \code{argsTTE}:
    \itemize{
    \item\code{CR} should competing risks be simulated? \cr 
    \item\code{rates.T} hazard corresponding to each endpoint (time to event endpoint, treatment group). \cr 
    \item\code{rates.C} same as \code{rates.T} but for the control group. \cr
    \item\code{rates.CR} same as \code{rates.T} but for the competing event (same in both groups). \cr
    \item\code{rates.Censoring} Censoring same as \code{rates.T} but for the censoring. \cr
    \item\code{name} names of the time to event variables. \cr
    \item\code{nameCensoring} names of the event type indicators. \cr
    }
}
\examples{

#### Providing p.1C and sHR ####
d <- simCompetingRisks(n.T = 5000, n.C = 5000, p.1C = 0.55, v.1C = -0.30, v.1T = -0.30, v.2C = -0.30, v.2T = -0.30, 
sHR = 0.5, b.1T = NULL, b.1C = NULL, b.2T = NULL, b.2C = NULL)

#### Providing the rate parameters ####
d <- simCompetingRisks(n.T=5000, n.C=5000, p.1C = NULL, v.1C = -0.30, v.1T = -0.30, v.2C = -0.30, v.2T = -0.30, 
sHR = NULL, b.1T = b.1T, b.1C = b.1C, b.2T = b.2T, b.2C = b.2C)

#### With exponential censoring ####
d <- simCompetingRisks(n.T = 5000, n.C = 5000, p.1C = 0.55, v.1C = -0.30, v.1T = -0.30, v.2C = -0.30, v.2T = -0.30, 
sHR = 0.5, b.1T = NULL, b.1C = NULL, b.2T = NULL, b.2C = NULL, cens.distrib = "exponential", param.cens = 0.8, 
latent = T)

### With uniform censoring ####
d <- simCompetingRisks(n.T = 5000, n.C = 5000, p.1C = 0.55, v.1C = -0.30, v.1T = -0.30, v.2C = -0.30, v.2T = -0.30, 
sHR = 0.5, b.1T = NULL, b.1C = NULL, b.2T = NULL, b.2C = NULL, cens.distrib = "uniform", param.cens = c(0, 7), 
latent=T)        

}
\keyword{function}
\keyword{simulations}
