% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.performance.R
\name{summary.performance}
\alias{summary.performance}
\title{Summary Method for Performance Objects}
\usage{
\method{summary}{performance}(object, order.model = NULL, digits = c(3, 3), print = TRUE, ...)
}
\arguments{
\item{object}{output of performance.}

\item{order.model}{[character vector] ordering of the models.}

\item{digits}{[numeric vector of length 2] number of digits used for the estimates and p-values.}

\item{print}{[logical] should the performance be printed in the console.}

\item{...}{not used.}
}
\description{
Summary of the performance of binary classifiers
}
