% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuyseTest-check.R
\name{testArgs}
\alias{testArgs}
\title{Check Arguments Passed to BuyseTest}
\usage{
testArgs(name.call, censoring, correction.uninf, cpus, data, endpoint,
  formula, iid, keep.pairScore, scoring.rule, model.tte, method.inference,
  n.resampling, hierarchical, neutral.as.uninf, operator, seed, strata,
  threshold, trace, treatment, type, weight, ...)
}
\description{
Check the validity of the argument passed the BuyseTest function by the user.
}
\keyword{internal}
