% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuyseRes-object.R
\docType{class}
\name{BuyseRes-class}
\alias{BuyseRes-class}
\alias{BuyseRes}
\title{Class "BuyseRes" (output of BuyseTest)}
\description{
A \code{\link{BuyseTest}} output is reported in a \code{BuyseRes} object.
}
\seealso{
\code{\link{BuyseTest}} for the function computing generalized pairwise comparisons. \cr
\code{\link{BuyseRes-summary}} for the summary of the BuyseTest function results
}
\keyword{BuyseRes-class}
\keyword{classes}
