% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/METHOD_get.R
\docType{methods}
\name{BuyseRes-getCount}
\alias{BuyseRes-getCount}
\alias{getCount}
\alias{getCount,BuyseRes-method}
\alias{show,getCount}
\title{get Method for Class "BuyseRes"}
\usage{
getCount(object, type)

\S4method{getCount}{BuyseRes}(object, type)
}
\arguments{
\item{object}{an \R object of class \code{\linkS4class{BuyseRes}}, i.e., output of \code{\link{BuyseTest}}}

\item{type}{the type of pairs to be counted. Can be \code{"favorable"}, \code{"unfavorable"}, \code{neutral}, or \code{uninf}. Can also be \code{"all"} to select all of them.}
}
\value{
A \code{"vector"} containing the number of pairs
}
\description{
Extract the number of pairs.
}
\examples{
dt <- simulBT(1e2)
BT <- BuyseTest(data=dt,endpoint="Y_TTE1",treatment="Treatment",
                type="timeToEvent",censoring="event1", n.bootstrap = 0)
getCount(BT)
getCount(BT, type = "favorable")

}
\keyword{BuyseRes-method}
\keyword{getCount}

