% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-setGeneric.R, R/BuyseRes_get.R
\docType{methods}
\name{getCount}
\alias{getCount}
\alias{BuyseRes-getCount}
\alias{getCount,BuyseRes-method}
\title{Extract the Number of Favorable, Unfavorable, Neutral, Uninformative pairs}
\usage{
getCount(object, type)

\S4method{getCount}{BuyseRes}(object, type)
}
\arguments{
\item{object}{an \R object of class \code{\linkS4class{BuyseRes}}, i.e., output of \code{\link{BuyseTest}}}

\item{type}{the type of pairs to be counted. Can be \code{"favorable"}, \code{"unfavorable"}, \code{neutral}, or \code{uninf}. Can also be \code{"all"} to select all of them.}
}
\value{
A \code{"vector"} containing the number of pairs
}
\description{
Extract the number of favorable, unfavorable, neutral, uninformative pairs.
}
\keyword{BuyseRes-method}
\keyword{get}
