% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-setGeneric.R, R/BuyseRes-iid.R
\docType{methods}
\name{iid}
\alias{iid}
\alias{BuyseRes-iid}
\alias{iid,BuyseRes-method}
\title{Extract the H-decomposition of the Estimator}
\usage{
iid(object, endpoint = NULL, order = 1)

\S4method{iid}{BuyseRes}(object, endpoint = NULL, order = 1)
}
\arguments{
\item{object}{an \R object of class \code{\linkS4class{BuyseRes}}, i.e., output of \code{\link{BuyseTest}}}

\item{endpoint}{[character] for which endpoint(s) the H-decomposition should be output?
If \code{NULL} returns the sum of the H-decomposition over all endpoints.}

\item{order}{[integer 1,2] which order of the H-decomposition should be output?
Can be the first or second order. \code{NULL} output a list containing all terms.}
}
\description{
Extract the H-decomposition of the GPC estimator.
}
\seealso{
\code{\link{BuyseTest}} for performing a generalized pairwise comparison. \cr
\code{\link{BuyseRes-summary}} for a more detailed presentation of the \code{BuyseRes} object.
}
\keyword{BuyseRes-method}
\keyword{iid}
