% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuyseRes_show.R
\docType{methods}
\name{BuyseRes-show}
\alias{BuyseRes-show}
\alias{show,BuyseRes-method}
\title{Show Method for Class "BuyseRes"}
\usage{
\S4method{show}{BuyseRes}(object)
}
\arguments{
\item{object}{an \R object of class \code{\linkS4class{BuyseRes}}, i.e., output of \code{\link{BuyseTest}}}
}
\description{
Display the main results stored in a \code{\link{BuyseRes}} object.
}
\seealso{
\code{\link{BuyseTest}} for performing a generalized pairwise comparison. \cr
  \code{\link{BuyseRes-summary}} for a more detailed presentation of the \code{BuyseRes} object.
}
\keyword{BuyseRes-method}
\keyword{summary}
