% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuyseRes-summary.R
\docType{methods}
\name{BuyseRes-summary}
\alias{BuyseRes-summary}
\alias{summary}
\alias{summary,BuyseRes}
\alias{summary,BuyseRes-method}
\title{Summary Method for Class "BuyseRes"}
\usage{
summary(object, ...)

\S4method{summary}{BuyseRes}(object, show = TRUE, percentage = TRUE,
  statistic = BuyseTest.options()$statistic, conf.level = 0.95,
  alternative = "two.sided", strata = if (length(object@level.strata) == 1)
  {     "global" } else {     NULL }, digit = c(2, 3))
}
\arguments{
\item{object}{output of \code{\link{BuyseTest}}}

\item{...}{arguments to be passed from the generic method to the class specific method [not relevant to the user]}

\item{show}{[logical] Should the table be displayed?.}

\item{percentage}{[logical] Should the percentage of pairs of each type be displayed ? Otherwise the number of pairs is displayed.}

\item{statistic}{[character] the statistic summarizing the pairwise comparison:
\code{"netChance"} displays the net chance in favor of treatment, as described in Buyse (2010) and Peron et al. (2016)),
whereas \code{"winRatio"} displays the win ratio, as described in Wang et al. (2016).}

\item{conf.level}{[numeric] confidence level for the confidence intervals.}

\item{alternative}{[character] the type of alternative hypothesis: \code{"two.sided"}, \code{"greater"}, or \code{"less"}.}

\item{strata}{[character vector] the name of the strata to be displayed. Can also be \code{"global"} to display the average over all strata.}

\item{digit}{[integer vector] the number of digit to use for printing the counts and the delta.}
}
\description{
Summarize the results from the \code{\link{BuyseTest}} function.
}
\details{
WARNING : when using a permutation test, the confidence interval is computed using quantiles of the distribution of the cumulative proportion in favor of the treatment
under the null hypothesis. 
It thus may not be valid if this hypothesis is rejected.

WARNING: For the win ratio, the proposed implementation enables the use of thresholds, endpoints that are not time to events,
and the correction proposed in Peron et al. (2016) to account for censoring. 
These development have not been examined by Wang et al. (2016), or in other papers (at out knowledge). They are only provided here by implementation convenience.
}
\examples{
dt <- simBuyseTest(1e2, n.strata = 3)

 \dontrun{
 BT <- BuyseTest(Treatment ~ TTE(eventtime, censoring = status) + Bin(toxicity), data=dt)
 }
 \dontshow{
 BT <- BuyseTest(Treatment ~ TTE(eventtime, censoring = status) + Bin(toxicity), data=dt, n.resampling = 10, trace = 0)
 }
 summary(BT)
 summary(BT, percentage = FALSE)
 summary(BT, statistic = "winRatio")

}
\seealso{
\code{\link{BuyseTest}} for performing a generalized pairwise comparison. \cr
  \code{\link{BuyseRes-class}} for a presentation of the \code{BuyseRes} object.
}
\keyword{BuyseRes-method}
\keyword{summary}
