% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuyseSim-summary.R
\docType{methods}
\name{BuyseSim-summary}
\alias{BuyseSim-summary}
\alias{summary,BuyseSim-method}
\title{Summary Method for Class "BuyseSim"}
\usage{
\S4method{summary}{BuyseSim}(object, print = TRUE, statistic = NULL,
  legend = TRUE, method.inference = NULL, col.rep = TRUE,
  digit = 4)
}
\arguments{
\item{object}{output of \code{\link{powerBuyseTest}}}

\item{print}{[logical] Should the table be displayed?.}

\item{statistic}{[character] the statistic summarizing the pairwise comparison:
\code{"netBenefit"} displays the net benefit
whereas \code{"winRatio"} displays the win ratio.}

\item{legend}{[logical] should explainations about the content of each column be displayed?}

\item{method.inference}{[character vector] for which inference method the rejection rate should be displayed?}

\item{col.rep}{[logical] should the number of successful simulations be displayed?}

\item{digit}{[integer vector] the number of digit to use for printing the counts and the delta.}

\item{...}{arguments to be passed from the generic method to the class specific method [not relevant to the user]}
}
\description{
Summarize the results from the \code{\link{powerBuyseTest}} function.
}
\seealso{
\code{\link{powerBuyseTest}} for performing a simulation study for generalized pairwise comparison. \cr
}
\keyword{BuyseSim-method}
\keyword{summary}
