% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuyseTest.options-object.R
\docType{class}
\name{BuyseTest.options-class}
\alias{BuyseTest.options-class}
\title{Class "BuyseTest.options" (global setting for the BuyseTest package)}
\description{
Class defining the global settings for the BuyseTest package.
}
\seealso{
\code{\link{BuyseTest.options}} to select or update global settings.
}
\keyword{BuyseTest.options-class}
\keyword{classes}
\keyword{options}
