% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCT_BuyseTest.R
\name{BuyseTest}
\alias{BuyseTest}
\title{Generalized Pairwise Comparisons}
\usage{
BuyseTest(formula, data, treatment = NULL, endpoint = NULL, type = NULL,
  threshold = NULL, censoring = NULL, strata = NULL,
  method = BuyseTest.options()$method,
  neutralAsUninf = BuyseTest.options()$neutralAsUninf,
  n.bootstrap = BuyseTest.options()$n.bootstrap, prob.alloc = NULL,
  stratified = FALSE, alternative = "two.sided",
  seed = BuyseTest.options()$seed, cpus = BuyseTest.options()$cpus,
  trace = BuyseTest.options()$trace)
}
\arguments{
\item{formula}{a symbolic description of the model to be fitted. The response variable should be a binary variable defining the treatment arms. 
The rest of the formula should indicate the strata variables (if any) and the endpoints by order of priority. \emph{formula}.}

\item{data}{A \code{data.frame} containing the variables.}

\item{treatment}{the name of the treatment variable identifying the control and the experimental group. \emph{character}.}

\item{endpoint}{the name of the endpoint variable(s). \emph{character vector}.}

\item{type}{the type of each endpoint. \emph{character vector}. Can be \code{"binary"}, \code{"continuous"} or \code{"timeToEvent"}.}

\item{threshold}{the thresholds, one for each endpoint variable. \emph{numeric vector}.}

\item{censoring}{the name of the censoring variable(s), one for each endpoint. \emph{character vector}.}

\item{strata}{the name of the strata variable(s). \emph{numeric vector}.}

\item{method}{Is defined when at least one time-to-event outcome is analyzed. Defines the method used to handle pairs which can not be decidely classified as favorable, unfavorable, or neutral because of censored observations.  Can be \code{"Gehan"}, \code{"Peto"}, \code{"Efron"}, or \code{"Peron"}. See details.}

\item{neutralAsUninf}{Should paired classified as neutral be re-analysed using endpoints of lower priority. \emph{logical}.}

\item{n.bootstrap}{the number of bootstrap samples used for computing the confidence interval and the p.values. \emph{integer}.}

\item{prob.alloc}{the resampling probability for assignement to the experimental group in the bootstrap samples. \emph{double}. Can also be \code{NULL} to use the proportion of patients in the experimental group.}

\item{stratified}{Should the boostrap be a stratified bootstrap? \emph{logical}.}

\item{alternative}{a \emph{character} specifying the alternative hypothesis. Must be one of \code{"two.sided"}, \code{"greater"} or \code{"less"}.}

\item{seed}{the seed to consider for the bootstrap. \emph{integer}.}

\item{cpus}{the number of CPU to use. \emph{integer}.}

\item{trace}{Should the execution of the function be traced ? \emph{integer}.}
}
\value{
An \R object of class \code{\linkS4class{BuyseRes}}.
}
\description{
Performs Generalized Pairwise Comparisons for binary, continuous and time-to-event outcomes.
}
\details{
\bold{treatment:} The variable corresponding to \code{treatment} in data must have only two levels (e.g. \code{0} and \code{1}). \cr
\bold{endpoint, threshold, censoring, and type:} Arguments \code{endpoint}, \code{threshold}, \code{censoring}  and \code{type} must have the same length. \cr
\code{threshold} must be \code{NA} for binary endpoints and positive for continuous or time to event endpoints. \cr
\code{censoring} must be \code{NA} for binary or continuous endpoints and indicate a variable in data for time to event endpoints. 
Short forms for endpoint \code{type} are \code{"bin"} (binary endpoint), \code{"cont"} (continuous endpoint), \code{"TTE"} (time-to-event endpoint). 

\bold{Bootstrap:} The number of bootstrap replications (argument \code{n.bootstrap}) must be specified to enable the computation of the confidence intervals and the p.value. 
A large number of bootstrap samples (e.g. \code{n.bootstrap=10000})  are needed to obtain accurate CI and p.value. See (Buyse et al., 2010) for more details. 

\bold{trace:} \code{3} reports all messages  \code{2} reports all messages except silent parallelization messages, \code{1} reports only the percentage of advancement of the bootstrap,  and \code{0} remains silent.

\bold{cpus parallelization:} Argument \code{cpus} can be set to \code{"all"} to use all available cpus. The parallelization relies on the \emph{snowfall} package (function \emph{sfClusterApplyLB}). The detection of the number of cpus relies on the \code{detectCores} function from the \emph{parallel} package .

\bold{Dealing with neutral or uninformative pairs:} Neutral pairs correspond to pairs for which the difference between the endpoint of the control observation and the endpoint of the treatment observation is (in absolute value) below the threshold. When \code{threshold=0}, neutral pairs correspond to pairs with equal outcome.\cr
Uninformative pairs correspond to pairs for which the censoring prevend from classifying them into favorable, unfavorable or neutral. Neutral or uninformative pairs for an endpoint with priority \code{l} are, when available, analysed on the endpoint with priority \code{l-1}.

\bold{method:} Pairs which can not be decidely classified as favorable, unfavorable, or neutral because of censored observations can be classified uninformative (\code{method="Gehan"}, Gehan 1965). 
Another solution is to estimate the probability for such pair to be classified as favorable, unfavorable, or neutral based on the survival functions. 
\code{method="Peto"} estimate these probabilities using the common Kaplan-Meier estimator of the survival function for treated and control patients (Peto et al. 1972). 
\code{method="Efron"}, and \code{method="Peron"} estimate these probabilities using separate Kaplan-Meier estimators of the survival functions for the two groups of patients. 
When the largest observation is censored, it is not possible to estimate the survival probability by the Kaplan-Meier estimator beyond this time point.  
\code{method="Efron"} treats the largest observations in each patient group as if it were uncensored (Efron 1967). 
\code{method="Peron"} treats the probability of survival beyond the last observation as NA, resulting in a non null probability that the pair is uninformative    
See Peron et al. (2016) for more details.
}
\examples{
# reset the default value of the number of bootstrap samples
BuyseTest.options(n.bootstrap = 0) # no bootstrap

#### simulate some data ####
df.data <- simulBT(1e2, n.strata = 2)

# display 
if(require(survival)){
resKM_tempo <- survfit(Surv(eventtime,status)~Treatment, data = df.data)
plot(resKM_tempo)
}

#### one time to event endpoint ####
BT <- BuyseTest(Treatment ~ TTE(eventtime, censoring = status), data=df.data)

summary(BT) # net chance in favor of treatment
summary(BT, percentage = FALSE)  
summary(BT, statistic = "winRatio") # win Ratio

## bootstrap to compute the CI
\dontrun{
BT <- BuyseTest(Treatment ~ TTE(eventtime, censoring = status), data=df.data,
                n.bootstrap = 1e3)
}
\dontshow{
  BT <- BuyseTest(Treatment ~ TTE(eventtime, censoring = status), data=df.data,
                  n.bootstrap = 1e1, trace = 0)
}
summary(BT)
summary(BT, statistic = "winRatio") # win Ratio

## parallel boostrap
\dontrun{
  BT <- BuyseTest(Treatment ~ TTE(eventtime, censoring = status), data=df.data,
                  n.bootstrap = 1e3, cpus = 2)
}
\dontshow{
  BT <- BuyseTest(Treatment ~ TTE(eventtime, censoring = status), data=df.data,
                  n.bootstrap = 1e1, cpus = 2, trace = 0)
}
summary(BT)

## method Gehan is much faster but does not optimally handle censored observations
\dontrun{
  BT <- BuyseTest(Treatment ~ TTE(eventtime, censoring = status), data=df.data,
                  n.bootstrap = 1e3, method = "Gehan")
}
\dontshow{
  BT <- BuyseTest(Treatment ~ TTE(eventtime, censoring = status), data=df.data,
                  n.bootstrap = 1e1, method = "Gehan", trace = 0)
}
summary(BT)

#### one time to event endpoint: only differences in survival over 1 count ####
BT <- BuyseTest(Treatment ~ TTE(eventtime, threshold = 1, censoring = status), data=df.data)
summary(BT)

#### one time to event endpoint with a strata variable
BT <- BuyseTest(Treatment ~ strata + TTE(eventtime, censoring = status), data=df.data)
summary(BT)

#### several endpoints with a strata variable
f <- Treatment ~ strata + T(eventtime, 1, status) + B(toxicity) 
f <- update(f, 
            ~. + T(eventtime, 0.5, status) + C(score, 1) + T(eventtime, 0.25, status))

BT <- BuyseTest(f, data=df.data)
summary(BT)

#### real example : Veteran dataset of the survival package ####
#### Only one endpoint. Type = Time-to-event. Thresold = 0. Stratfication by histological subtype
#### method = "Gehan"

if(require(survival)){
\dontrun{
  data(veteran,package="survival")
  n.bootstrap <- 100
 
  ## method = "Gehan"
  BT_Gehan <- BuyseTest(trt ~ celltype + TTE(time,threshold=0,censoring=status), 
                        data=veteran, n.bootstrap=n.bootstrap,method="Gehan")
  
  summary_Gehan <- summary(BT_Gehan)
  summary_Gehan <- summary(BT_Gehan, statistic = "winRatio")
  
  ## method = "Peron"
  BT_Peron <- BuyseTest(data=veteran,endpoint="time",treatment="trt",strata="celltype",
                        type="timeToEvent",censoring="status",threshold=0,
                        n.bootstrap=n.bootstrap,method="Peron")
  
  summary_Peron <- summary(BT_Peron)
}
}
}
\references{
Marc Buyse (2010). \bold{Generalized pairwise comparisons of prioritized endpoints in the two-sample problem}. \emph{Statistics in Medicine} 29:3245-3257 \cr
D. Wang, S. Pocock (2016). \bold{A win ratio approach to comparing continuous non-normal outcomes in clincal trials}. \emph{Pharmaceutical Statistics} 15:238-245 \cr
J. Peron, M. Buyse, B. Ozenne, L. Roche and P. Roy (2016). \bold{An extension of generalized pairwise comparisons for prioritized outcomes in the presence of censoring}. Statistical Methods in Medical Research. \cr
Efron B (1967). \bold{The two sample problem with censored data}. \emph{Proceedings of the Fifth Berkeley Symposium on Mathematical Statistics and Probability} 4:831-583 \cr
Peto R, Peto J (1972). \bold{Asymptotically efficient rank invariant test procedures}. \emph{Journal of the Royal Statistical Society - series A} 135(2):185-198 \cr
Gehan EA (1965). \bold{A generalized two-sample Wilcoxon test for doubly censored data}. \emph{Biometrika}  52(3):650-653 \cr
}
\seealso{
\code{\link{BuyseRes-summary}} for a summary of the results of generalized pairwise comparison. \cr
\code{\link{BuyseRes-class}} for a presentation of the \code{BuyseRes} object. \cr
\code{\link{constStrata}} to create a strata variable from several clinical variables. \cr
}
\keyword{BuyseTest}
\keyword{function}

