% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuyseTest-print.R
\name{internal-print}
\alias{internal-print}
\alias{printGeneral}
\alias{printInference}
\title{internal functions for BuyseTest - display}
\usage{
printGeneral(
  status,
  D,
  D.TTE,
  data,
  endpoint,
  hierarchical,
  level.strata,
  level.treatment,
  scoring.rule,
  M.status,
  neutral.as.uninf,
  correction.uninf,
  operator,
  strata,
  threshold,
  trace,
  treatment,
  type,
  weight,
  Wscheme,
  ...
)

printInference(method.inference, n.resampling, cpus, seed, ...)
}
\description{
Functions called by \code{\link{BuyseTest}} to display the settings.
}
\author{
Brice Ozenne
}
\keyword{internal}
