% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuyseTest-print.R
\name{internal-print}
\alias{internal-print}
\alias{printGeneral}
\alias{printPermutation}
\title{internal functions for BuyseTest - display}
\usage{
printGeneral(levels.treatment, levels.strata, n.strata, endpoint, threshold,
  censoring, operator, type, D, D.TTE, method, neutralAsUninf, Wscheme,
  threshold_TTEM1)
}
\description{
Functions called by \code{\link{BuyseTest}} to display what is going on.
}
\details{
\code{printGeneral} display general settings. \cr
\code{printPermutation} display the settings of the permutation test. \cr
}
\keyword{BuyseTest}
\keyword{function}
\keyword{internal}
