% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuyseTest-print.R
\name{internal-print}
\alias{internal-print}
\alias{printGeneral}
\alias{printInference}
\title{internal functions for BuyseTest - display}
\usage{
printGeneral(censoring, D, D.TTE, data, endpoint, level.strata,
  level.treatment, method.tte, neutral.as.uninf, correction.uninf,
  operator, strata, threshold, trace, treatment, type, Wscheme, ...)

printInference(method.inference, n.resampling, cpus, seed, ...)
}
\description{
Functions called by \code{\link{BuyseTest}} to display the settings.
}
\keyword{internal}
