% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCTi-print.R
\name{internal-print}
\alias{internal-print}
\alias{printBoostrap}
\alias{printGeneral}
\title{internal functions for BuyseTest - display}
\usage{
printGeneral(levels.treatment, levels.strata, n.strata, endpoint, threshold,
  censoring, type, D, D.TTE, method, Wscheme, threshold_TTEM1)
}
\description{
Functions called by \code{\link{BuyseTest}} to display what is going on.
}
\details{
\code{printGeneral} display general settings. \cr
\code{printBoostrap} display bootstrap settings. \cr
}
\keyword{BuyseTest}
\keyword{function}
\keyword{internal}

