% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-BuysePower-summary.R
\docType{methods}
\name{S4BuysePower-summary}
\alias{S4BuysePower-summary}
\alias{summary,S4BuysePower-method}
\title{Summary Method for Class "S4BuysePower"}
\usage{
\S4method{summary}{S4BuysePower}(object, print = TRUE, legend = TRUE, col.rep = FALSE, digit = 4, ...)
}
\arguments{
\item{object}{output of \code{\link{powerBuyseTest}}}

\item{print}{[logical] Should the table be displayed?.}

\item{legend}{[logical] should explainations about the content of each column be displayed?}

\item{col.rep}{[logical] should the number of successful simulations be displayed?}

\item{digit}{[integer vector] the number of digit to use for printing the counts and the delta.}

\item{...}{arguments passed to the method \code{model.tables} (\code{statistic}, \code{endpoint}, \code{order.Hprojection}, \code{transformation})}
}
\value{
data.frame
}
\description{
Summarize the results from the \code{\link{powerBuyseTest}} function.
}
\seealso{
\code{\link{powerBuyseTest}} for performing a simulation study for generalized pairwise comparison. \cr
}
\author{
Brice Ozenne
}
\keyword{print}
