% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-BuyseTest-nobs.R
\docType{methods}
\name{S4BuyseTest-nobs}
\alias{S4BuyseTest-nobs}
\alias{nobs,S4BuyseTest-method}
\title{Sample Size for Class "S4BuyseTest"}
\usage{
\S4method{nobs}{S4BuyseTest}(object, stratified = NULL, ...)
}
\arguments{
\item{object}{an \R object of class \code{S4BuyseTest}, i.e., output of \code{\link{BuyseTest}}}

\item{stratified}{[logical] should strata-specific sample size be output?
Otherwise output the sample size and number of pairs over all strata.}

\item{...}{no used, for compatibility with the generic method.}
}
\value{
A vector (when argument \code{stratified} is \code{FALSE}) or a data.frame (when argument \code{stratified} is \code{TRUE}). In the latter case each line correspond to a strata.
}
\description{
Display the sample size in each treatmnet arm as well as the number of pairs.
}
\author{
Brice Ozenne
}
\keyword{methods}
