% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-BuyseTest-autoplot.R
\docType{methods}
\name{S4BuyseTest-plot}
\alias{S4BuyseTest-plot}
\alias{plot,S4BuyseTest-method}
\alias{autoplot,S4BuyseTest-method}
\alias{plot,S4BuyseTest,ANY-method}
\title{Graphical Display for GPC}
\usage{
\S4method{autoplot}{S4BuyseTest}(
  object,
  type = "hist",
  strata = "global",
  endpoint = NULL,
  label.strata = NULL,
  label.endpoint = NULL,
  size.text = 5,
  color = c("#7CAE00", "#F8766D", "#C77CFF", "#00BFC4"),
  ...
)

\S4method{plot}{S4BuyseTest,ANY}(x, plot = TRUE, ...)
}
\arguments{
\item{object, x}{an \R object of class \code{\linkS4class{S4BuyseTest}}, i.e., output of \code{\link{BuyseTest}}}

\item{type}{[character] type of plot: histogram (\code{"hist"}), pie chart (\code{"pie"}), or nested pie charts (\code{"racetrack"}).}

\item{strata}{[character vector] strata(s) relative to which the percentage should be displayed.}

\item{endpoint}{[character vector] endpoint(s) relative to which the percentage should be displayed.}

\item{label.strata}{[character vector] new labels for the strata levels. Should match the length of argument \code{strata}.}

\item{label.endpoint}{[character vector] new labels for the endpoints. Should match the length of argument \code{endpoint}.}

\item{size.text}{[numeric, >0] size of the text indicating the endpoints in the nested pie charts.}

\item{color}{[character vector] colors used to display the percentages for each type of pair.}

\item{...}{either not used (\code{autoplot}, for compatibility with the generic function) or passed to \code{autoplot} (\code{plot}).}

\item{plot}{[logical] should the graphic be displayed in a graphical window.}
}
\description{
Graphical display of the percentage of favorable, unfavorable, neutral, and uninformative pairs.
}
\details{
The \code{autoplot} and \code{plot} methods are very similar. The main difference is that the former returns a ggplot2 object whereas the later automatically display the figure in a graphical window and returns an (invible) list with the plot and the data.
}
\examples{
\dontrun{
require(ggplot2)

## simulate data
set.seed(10)
df.data <- simBuyseTest(1e2, n.strata = 2)

ff1 <- treatment ~ bin(toxicity) + TTE(eventtime, status = status,
                                      restriction = 1, threshold = 0.5)
BT1 <- BuyseTest(ff1, data= df.data)
plot(BT1, type = "hist")
plot(BT1, type = "pie")
plot(BT1, type = "racetrack")

ff2 <- update(ff1, ~.+cont(score))
BT2 <- BuyseTest(ff2, data= df.data)
plot(BT2, type = "hist")
plot(BT2, type = "pie")
plot(BT2, type = "racetrack")
}
}
