% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-BuyseTest-show.R
\docType{methods}
\name{S4BuyseTest-show}
\alias{S4BuyseTest-show}
\alias{show,S4BuyseTest-method}
\title{Show Method for Class "S4BuyseTest"}
\usage{
\S4method{show}{S4BuyseTest}(object)
}
\arguments{
\item{object}{an \R object of class \code{S4BuyseTest}, i.e., output of \code{\link{BuyseTest}}}
}
\description{
Display the main results stored in a \code{S4BuyseTest} object.
}
\seealso{
\code{\link{BuyseTest}} for performing a generalized pairwise comparison. \cr
  \code{\link{S4BuyseTest-summary}} for a more detailed presentation of the \code{S4BuyseTest} object.
}
\author{
Brice Ozenne
}
\keyword{S4BuyseTest-method}
\keyword{summary}
