% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-BuyseTest-update.R
\docType{methods}
\name{S4BuyseTest-update}
\alias{S4BuyseTest-update}
\alias{update,S4BuyseTest-method}
\title{Re-run two-group GPC}
\usage{
\S4method{update}{S4BuyseTest}(object, ...)
}
\arguments{
\item{object}{an \R object of class \code{S4BuyseTest}, i.e., output of \code{\link{BuyseTest}}}

\item{...}{additional arguments to the call, or arguments with changed values.}
}
\value{
an \R object of class \code{S4BuyseTest}
}
\description{
Re-run Generalized Pairwise Comparisons (GPC) between two groups.
}
\author{
Brice Ozenne
}
\keyword{methods}
