% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuyseTest-check.R
\name{testArgs}
\alias{testArgs}
\title{Check Arguments Passed to BuyseTest}
\usage{
testArgs(
  name.call,
  status,
  correction.uninf,
  cpus,
  data,
  endpoint,
  engine,
  formula,
  iid,
  iidNuisance,
  keep.pairScore,
  scoring.rule,
  pool.strata,
  model.tte,
  method.inference,
  n.resampling,
  strata.resampling,
  hierarchical,
  neutral.as.uninf,
  add.halfNeutral,
  operator,
  censoring,
  restriction,
  seed,
  strata,
  threshold,
  trace,
  treatment,
  type,
  weightEndpoint,
  weightObs,
  ...
)
}
\description{
Check the validity of the argument passed the BuyseTest function by the user.
}
\author{
Brice Ozenne
}
\keyword{internal}
