% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuyseTest-check.R
\name{testArgs}
\alias{testArgs}
\title{Check Arguments Passed to BuyseTest}
\usage{
testArgs(alternative, name.call, censoring, correction.uninf, cpus, data,
  endpoint, formula, keep.pairScore, method.tte, model.tte,
  method.inference, n.resampling, neutral.as.uninf, operator, seed, strata,
  threshold, trace, treatment, type, ...)
}
\description{
Check the validity of the argument passed the BuyseTest function by the user.
}
\keyword{internal}
